Time: 1 ms (99.98%) | Memory: 42.8 MB (50.05%) - LeetSync

+++ 4303-count-k-th-roots-in-a-range/count-k-th-roots-in-a-range.java
@@ -4,10 +4,10 @@ class Solution {
             return r - l + 1;
         }
         int count = 0;
-        for(long i=0;i<=r;i++){
+        for(int i=0;i<=r;i++){
             long power = (long)Math.pow(i, k);
             if(power>r) break;
-            if(power>=l&&power<=r){
+            if(power>=l && power<=r){
                 count++;
             }
         }
